# week_7/library_app/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Library API
  version: "1.0.0"
  description: Library catalogue with session-based auth and role-based access.
servers:
  - url: /
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: library.sid
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
    Book:
      type: object
      properties:
        _id: { type: string }
        title: { type: string }
        author: { type: string }
        isbn: { type: string }
        genre: { type: string }
        publishedYear: { type: integer }
        copiesTotal: { type: integer }
        copiesAvailable: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            details:
              type: array
              items: { type: string }
paths:
  /api/books:
    get:
      summary: List books
      parameters:
        - in: query
          name: author
          schema: { type: string }
        - in: query
          name: genre
          schema: { type: string }
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Book" }
        "400":
          description: Invalid genre
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      summary: Create a book (librarian only)
      security:
        - cookieAuth: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Book" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Book" }
        "400": { description: Validation failed, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { description: Not logged in }
        "403": { description: Not a librarian, or missing CSRF token }
        "409": { description: Duplicate ISBN }
  /api/books/{id}:
    get:
      summary: Get one book
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Book" } } } }
        "400": { description: Invalid id }
        "404": { description: Not found }
    patch:
      summary: Update a book (librarian only)
      security:
        - cookieAuth: []
          csrfHeader: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Book" }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/Book" } } } }
        "400": { description: Validation failed }
        "401": { description: Not logged in }
        "403": { description: Not a librarian, or missing CSRF token }
        "404": { description: Not found }
    delete:
      summary: Delete a book (librarian only)
      security:
        - cookieAuth: []
          csrfHeader: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
        "401": { description: Not logged in }
        "403": { description: Not a librarian, or missing CSRF token }
        "404": { description: Not found }
...